requirements.txt
discord.py>=2.6.4
openai>=1.0.0
requests